// src/types/user/action.types.ts
export enum AuthActions {
  LOGOUT = 'LOGOUT',
  RESET_USERS = 'RESET_USERS',
  RESET_USER_DATA = 'RESET_USER_DATA',
  SET_ACCESS_TOKEN = 'SET_ACCESS_TOKEN',
  SET_USERS = 'SET_USERS',
  SET_USER_DATA = 'SET_USER_DATA',
  SET_USER_IS_LOGGED = 'SET_USER_IS_LOGGED',
  USER_HAS_TO_UPDATE = 'USER_HAS_TO_UPDATE'
}